requirements.txt
# Web scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Document processing
pdfplumber==0.11.0
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# LangChain and AI
langchain==0.1.9
langchain-community==0.0.24
sentence-transformers==2.3.1
chromadb==0.4.22
faiss-cpu==1.7.4

# Web framework
flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.1
openai==1.12.0
tiktoken==0.6.0
